refactor: move msw from shared/lib to shared/api
