+++ src/features/article/favorite/favorite-article/api/favoriteArticles.test.ts
@@ -2,8 +2,8 @@ import { QueryClient } from '@tanstack/react-query';
 import { act, renderHook } from '@testing-library/react';
 import { vi } from 'vitest';
 import { articleApi } from '~entities/article';
+import { wait } from '~shared/api/msw';
 import { ArticleDto, realworldApi } from '~shared/api/realworld';
-import { wait } from '~shared/lib/msw';
 import { createWrapper } from '~shared/lib/react-query';
 import { useMutationFavoriteArticle } from './favoriteArticle';
 import { setupPostFavoriteArticleHandlers } from './msw/postFavoriteArticleHandlers';
