+++ src/features/article/favorite/unfavorite-article/api/unfavoriteArticle.test.ts
@@ -3,8 +3,8 @@ import { renderHook } from '@testing-library/react';
 import { act } from 'react-dom/test-utils';
 import { vi } from 'vitest';
 import { articleApi } from '~entities/article';
+import { wait } from '~shared/api/msw';
 import { ArticleDto, realworldApi } from '~shared/api/realworld';
-import { wait } from '~shared/lib/msw';
 import { createWrapper } from '~shared/lib/react-query';
 import { setupDeleteUnfavoriteArticleHandlers } from './msw/deleteUnfavoriteArticleHandlers';
 import { useMutationUnfavoriteArticle } from './unfavoriteArticle';
