+++ src/features/article/update/api/msw/putUpdateArticleHandlers.ts
@@ -1,11 +1,11 @@
 import { rest } from 'msw';
-import { UpdateArticleDto, realworldApi } from '~shared/api/realworld';
 import {
   server,
   initTestDatabase,
   parseTokenFromRequest,
   mapMswArticleDto,
-} from '~shared/lib/msw';
+} from '~shared/api/msw';
+import { UpdateArticleDto, realworldApi } from '~shared/api/realworld';
 
 const databaseApi = initTestDatabase();
