+++ src/features/comment/create/api/createComment.test.ts
@@ -3,8 +3,8 @@ import { renderHook } from '@testing-library/react';
 import { act } from 'react-dom/test-utils';
 import { vi } from 'vitest';
 import { commentApi } from '~entities/comment';
+import { wait } from '~shared/api/msw';
 import { CommentDto, realworldApi } from '~shared/api/realworld';
-import { wait } from '~shared/lib/msw';
 import { createWrapper } from '~shared/lib/react-query';
 import { useCreateComment } from './createComment';
 import { setupPostCreateArticleHandlers } from './msw/postCreateCommentHandlers';
