+++ src/features/comment/delete/api/deleteComment.test.ts
@@ -2,8 +2,8 @@ import { QueryClient } from '@tanstack/react-query';
 import { act, renderHook } from '@testing-library/react';
 import { vi } from 'vitest';
 import { commentApi } from '~entities/comment';
+import { wait } from '~shared/api/msw';
 import { CommentDto, realworldApi } from '~shared/api/realworld';
-import { wait } from '~shared/lib/msw';
 import { createWrapper } from '~shared/lib/react-query';
 import { useDeleteComment } from './deleteComment';
 import { setupdeleteCommentHandlers } from './msw/deleteCommentHandlers';
