+++ src/features/profile/follow/follow-profile/api/followUser.test.ts
@@ -2,8 +2,8 @@ import { QueryClient } from '@tanstack/react-query';
 import { act, renderHook } from '@testing-library/react';
 import { vi } from 'vitest';
 import { profileApi } from '~entities/profile';
+import { wait } from '~shared/api/msw';
 import { ProfileDto, realworldApi } from '~shared/api/realworld';
-import { wait } from '~shared/lib/msw';
 import { createWrapper } from '~shared/lib/react-query';
 import { useMutationFollowUser } from './followUser';
 import { setupPostFollowUserHandlers } from './msw/postFollowUserHandlers';
