+++ src/shared/api/msw/lib/parseTokenFromRequest.ts
